+++ apps/api/src/env.ts
@@ -14,6 +14,8 @@ export const env = createEnv({
     STRIPE_WEBHOOK_SECRET: z.string().min(1),
     OPENROUTER_API_KEY: z.string().min(1),
     DEEPGRAM_API_KEY: z.string().min(1),
+    ASSEMBLYAI_API_KEY: z.string().min(1).optional(),
+    SONIOX_API_KEY: z.string().min(1).optional(),
   },
   runtimeEnv: Bun.env,
   emptyStringAsUndefined: true,
